+++ src/DataCollectors/DataCollectorInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Support\DataCollectors;
+
+use Illuminate\Http\Request;
+
+interface DataCollectorInterface
+{
+    public function collect(Request $request);
+}